Make all the tests pass
- component is almost certainly not working correctly yet, but at least stop it from breaking all the other tests
- adds a placeholder spec file for the component template, will be filled in later
