+++ app/assets/javascripts/govuk_publishing_components/lib/global-bar-init.js
