+++ app/views/govuk_publishing_components/components/_global_bar.html.erb
@@ -18,9 +18,9 @@
   # Regardless of value, banner is always manually dismissable by users
   always_visible = true
 
-  global_bar_classes = %w(global-bar govuk-!-display-none-print)
+  global_bar_classes = %w(gem-c-global-bar govuk-!-display-none-print)
 
-  title_classes = %w(global-bar-title)
+  title_classes = %w(gem-c-global-bar-title)
   title_classes << "js-call-to-action" if title_href
   title_classes << "govuk-link" if title_href
 
@@ -35,7 +35,7 @@
 <% if show_global_bar %>
   <!--[if gt IE 7]><!-->
   <div id="global-bar" class="<%= global_bar_classes.join(' ') %>" data-module="global-bar" <%= "data-global-bar-permanent=true" if always_visible %> data-nosnippet>
-    <p class="global-bar-message govuk-width-container">
+    <p class="gem-c-global-bar-message govuk-width-container">
       <% if title %>
         <% if title_href %>
           <a class="<%= title_classes.join(' ') %>" href="<%= title_href %>" data-module="ga4-link-tracker" data-ga4-link="<%= ga4_data %>"><%= title %></a>
@@ -45,7 +45,7 @@
       <% end %>
 
       <% if link_text %>
-        <span class="global-bar-text">
+        <span class="gem-c-global-bar-text">
         <% if link_href %>
           <%= link_to(
             link_text,
