+++ spec/components/global_bar_spec.rb
@@ -1,0 +1,7 @@
+require "rails_helper"
+
+describe "Global bar", type: :view do
+  def component_name
+    "global_bar"
+  end
+end
